:sparkles: Add 'utils' folder in src for extra Python tools
This commit introduces a new folver called 'utils'within the 'src' directory. The 'utils' folder will be used to store several extra Python tools that are not directly related to the main modules of the project. This helps with code organization an maintenance, making it easier to find additional utilities when needed

The following files were added to the 'utils' folder:

- cnpj.py: Formats numbers to the CNPJ standard and corrects incorrect CNPJ.

- date.py: Provides functions for formatting dates.

- hour.py: Offers functionalities related to the hour format.

- ler.py: Implements functionalities for reading data.

These tools are designed to assist in formatting spreadsheets and provide functionality related to CNPJ, dates, times and data reading.